refactor: Util 관련 클래스 support/util 이동
