+++ src/main/java/net/detalk/api/admin/service/VisitorLogService.java
@@ -12,9 +12,9 @@ import net.detalk.api.admin.controller.response.VisitorLogResponse;
 import net.detalk.api.admin.domain.VisitorLog;
 import net.detalk.api.admin.domain.exception.VisitorLocationSaveException;
 import net.detalk.api.admin.repository.VisitorLogRepository;
-import net.detalk.api.support.EnvironmentHolder;
+import net.detalk.api.support.util.EnvironmentHolder;
 import net.detalk.api.support.paging.PagingData;
-import net.detalk.api.support.TimeHolder;
+import net.detalk.api.support.util.TimeHolder;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.scheduling.annotation.Async;
