+++ src/main/java/net/detalk/api/image/service/ImageService.java
@@ -6,8 +6,8 @@ import net.detalk.api.image.domain.AttachmentFile;
 import net.detalk.api.image.controller.response.UploadImageData;
 import net.detalk.api.image.domain.exception.InvalidImageFormatException;
 import net.detalk.api.image.repository.AttachmentFileRepository;
-import net.detalk.api.support.TimeHolder;
-import net.detalk.api.support.UUIDGenerator;
+import net.detalk.api.support.util.TimeHolder;
+import net.detalk.api.support.util.UUIDGenerator;
 import net.detalk.api.support.image.ImageClient;
 import net.detalk.api.support.image.UploadImageInfo;
 import org.springframework.stereotype.Service;
