+++ src/main/java/net/detalk/api/infrastructure/bootstrap/DetalkInitializer.java
@@ -12,7 +12,7 @@ import net.detalk.api.alarm.service.AlarmSender;
 import net.detalk.api.role.domain.Role;
 import net.detalk.api.plan.service.PricingPlanCache;
 import net.detalk.api.role.repository.RoleRepository;
-import net.detalk.api.support.EnvironmentHolder;
+import net.detalk.api.support.util.EnvironmentHolder;
 import net.detalk.api.support.security.SecurityRole;
 import org.springframework.boot.ApplicationArguments;
 import org.springframework.boot.ApplicationRunner;
