+++ src/main/java/net/detalk/api/member/service/MemberService.java
@@ -13,8 +13,8 @@ import net.detalk.api.member.domain.exception.MemberProfileNotFoundException;
 import net.detalk.api.domain.exception.UserHandleDuplicatedException;
 import net.detalk.api.member.repository.MemberProfileRepository;
 import net.detalk.api.member.repository.MemberRepository;
-import net.detalk.api.support.TimeHolder;
-import net.detalk.api.support.UUIDGenerator;
+import net.detalk.api.support.util.TimeHolder;
+import net.detalk.api.support.util.UUIDGenerator;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
