+++ src/main/java/net/detalk/api/support/filter/MDCFilter.java
@@ -6,7 +6,7 @@ import jakarta.servlet.ServletException;
 import jakarta.servlet.ServletRequest;
 import jakarta.servlet.ServletResponse;
 import jakarta.servlet.http.HttpServletRequest;
-import net.detalk.api.support.UUIDGenerator;
+import net.detalk.api.support.util.UUIDGenerator;
 import org.slf4j.MDC;
 import org.springframework.core.annotation.Order;
 import org.springframework.stereotype.Component;
