+++ src/main/java/net/detalk/api/support/util/Constant.java
@@ -1,4 +1,4 @@
-package net.detalk.api.support;
+package net.detalk.api.support.util;
 
 public class Constant {
     public static final String COOKIE_ACCESS_TOKEN = "at";
