+++ src/main/java/net/detalk/api/support/util/InstantConverter.java
@@ -1,4 +1,4 @@
-package net.detalk.api.support;
+package net.detalk.api.support.util;
 
 import org.jooq.Converter;
