+++ src/main/java/net/detalk/api/support/util/TimeHolder.java
@@ -1,4 +1,4 @@
-package net.detalk.api.support;
+package net.detalk.api.support.util;
 
 import java.time.Instant;
 import java.time.LocalDateTime;
