+++ src/test/java/net/detalk/api/service/ProductPostServiceTest.java
@@ -40,8 +40,8 @@ import net.detalk.api.product.repository.ProductRepository;
 import net.detalk.api.post.service.ProductPostIdempotentService;
 import net.detalk.api.post.service.ProductPostService;
 import net.detalk.api.support.paging.CursorPageData;
-import net.detalk.api.support.TimeHolder;
-import net.detalk.api.support.UUIDGenerator;
+import net.detalk.api.support.util.TimeHolder;
+import net.detalk.api.support.util.UUIDGenerator;
 import net.detalk.api.support.error.ApiException;
 import net.detalk.api.tag.service.TagService;
 import org.junit.jupiter.api.BeforeEach;
