+++ src/test/java/net/detalk/api/support/security/TokenProviderTest.java
@@ -4,7 +4,7 @@ import java.util.List;
 import java.util.UUID;
 import net.detalk.api.mock.FakeUUIDGenerator;
 import net.detalk.api.support.config.AppProperties;
-import net.detalk.api.support.UUIDGenerator;
+import net.detalk.api.support.util.UUIDGenerator;
 import net.detalk.api.support.error.ExpiredTokenException;
 import net.detalk.api.support.error.TokenException;
 import org.junit.jupiter.api.BeforeEach;
